backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
openai==1.30.1
chromadb==0.5.0
rank_bm25==0.2.2
PyPDF2==3.0.1
python-docx==1.1.2
markdown==3.6
tiktoken==0.7.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
httpx==0.27.0
redis==5.0.4
numpy<2
greenlet>=3,<4
